Security Issue when switching into another club while logged in
Security issue - Login sessions cross over to multiple clubs.   An
attacker would have to be a club member of one of the clubs Synergy
supports.  If he/she knew of the other clubs, the user could inject a
simple get request to another club, and easily add a match against a
member from another club.