Display "lost" reservations
A reservation is created with an hours duration set, which means that
the reservation is of a different duration than the courts default.
When this occurs the time slots available after that are adjusted by
that difference.  The problem is once that reservation with the hours
duration set is removed, then the reservations later in the day are
made with an offset and consequently never displayed.